Localisation
Setted up all localisation
Ability to change translated value
Ability to create achievements and prizes in en and ru
Choose currency when creating a wish

+++ getmestuff/database/migrations/2017_06_04_110037_create_wishes_table.php
@@ -18,8 +18,8 @@ class CreateWishesTable extends Migration
             $table->integer('user_id')->unsigned();
             $table->string('url', 255);
             $table->json('address');
-            $table->integer('amount_needed')->unsigned();
-            $table->integer('current_amount')->unsigned()->default(0);
+            $table->float('amount_needed')->unsigned();
+            $table->float('current_amount')->unsigned()->default(0);
             $table->tinyInteger('priority')->unsigned()->default(1);
             $table->boolean('validated')->unsigned()->default(1);
             $table->boolean('completed')->unsigned()->default(0);
